Change: Release biglock when calling fuse_internal_vnode_disappear. This also requires fuse_vnop_strategy to hold the biglock, so this change is included.
The reason that we want to do this is that fuse_internal_vnode_disappear can reclaim the vnode, passing the request on to other file systems or possibly this one again! This can create a deadlock, and we want to avoid that.
Long term, we want to be able to convert biglock to a regular mutex and do away with the recursive lock, and this is one piece of that puzzle.