Fix deadlock issue on 64 bit kernel as seen with EncFS under heavy I/O
Release biglock before calling "int msleep(...)" and reacquire it after wakeup.
The kernel extension calls msleep in function "int fticket_wait_answer(...)"
when waiting for an response from the user space daemon. This can lead to a
deadlock, if the user space daemon inadvertently causes an operation on the
FUSE file system.

Based on the findings of Anatol Pomozov