Fix a file system deadlock in fuse_vnop_open
The deadlock is triggered in fuse_vnop_open, if one threads tries to open a
file another thread was just about to create. The thread wanting to open the
file goes to sleep, waiting for the other thread to create the file but never
releases the node lock. As a result the creating thread cannot acquire the
lock and is doomed to wait forever. Deadlock.