Use 64-bit locking code under 32-bit kernels
This is done to address a race condition, in the implementation of
fuse_vnop_create/fuse_vnop_open in combination with the 32-bit kernel's
fallback locking mechanism for non-thread-safe file systems, that can lead to a
file system deadlock. This issue affected 64-bit kernels as well and has been
fixed in commit ea1ada3296f29202a1939f84e8c73a172f9e3743.

We assume, that a call to fuse_vnop_create is always followed by a call to
fuse_vnop_open by the same thread. Once fuse_vnop_create returns, the vnode of
the new file is accessible in subsequent fuse_vnop_lookup calls. This allows
contenders to look up the vnode and try to open the file between the call to
fuse_vnop_create and fuse_vnop_open. Contenders are prevented from completing
the call to fuse_vnop_open as long as the flag FN_CREATING is set and are send
to sleep without releasing the vnode lock of the file. Since locking is done
by the kernel, we cannot release the vnode lock.