Implement fuse_kludge_vnode_isrecycled.
This is a 'hack' which implements vnode_isrecycled for systems where
this function is not exported by the kernel (< Mac OS X 10.6).

It is implemented in a similar way to how the kernel implements it:
1. Lock the vnode.
2. Read v_lflag.
3. Unlock the vnode.

struct vnode is normally an opaque struct with unknown definition, so we
use a prototype that is good enough to extract the relevant fields
(v_lock and v_lflag).