mount_osxfusefs: Fix 64-bit issue in option parsing.
Option parsing worked by passing a generic pointer as argument to
type-specific parse functions. However when the target was not a pointer
but a uint32_t variable, the parsing routine would for 64-bit platforms
write 8 bytes to the location anyway. This means that adjacent data
would be overwritten and in the case of big-endian architectures the
value would be lost.

Fixed by changing the types of variables that are passed as pointers to
the generic parsing function into uintptr_t. This means that they are
exactly as large as a pointer and no further problems should arise.