Actively wait for user space daemon to complete FUSE_INIT handshake
If the FUSE_INIT handshake is not completed by the user space daemon within
deamon_timeout mark the file system as dead. This makes sure processes
accessing an uninitialized FUSE file system are not being blocked indefinitely
while waiting for the potentially dead user space daemon to complete the
handshake.