Fix fuse_vnop_write/fuse_vnop_lookup race condition
Do not alter a looked up node's filesize or attributes. Doing so can
cause data corruption:

If a file vnode is being looked up while another thread is extending
the file, fuse_vnop_lookup might override the the file's new extended
size. This in turn causes fuse_vnop_blockmap to return a negative or
zero "run" value, when being called by custer_io. As a result the write
operation either fails with EINVAL or the file's content is corrupted.