Fixes hang on truncated direct writes.
Writes should stop when the filesystem refuses to accept additional
data, rather than looping infinitely.  This is consistent with
reads and with the Linux fuse behavior.

In this particular case, the looping was actually within uiomove(),
due to the inconsistent update of the uio, though other iov
configurations might have misbehaved differently.