Write back changes when unmapping file
munmap(2) states:

  If the mapping maps data from a file (MAP_SHARED), then the memory
  will eventually be written back to disk if it's dirty. This will
  happen automatically at some point in the future (implementation
  dependent).

In our case this point is now. Maintaining a consistent file state on
the backing storage is important if we are dealing with distributed
file systems.

Note:

fuse_vnop_mnomap() is called when there are no more references to the
file's mapped data. In other words, we have no definate way of knowing
when a particular process unmaps a file since the file's data might
still be referenced by other processes (MAP_SHARED).

Calling ubc_msync() in fuse_vnop_close() is not going to help because
the close(2) system call does not unmap the file. See mmap(2).

Since vnop mnomap is considered a hint, returned errors are being
ignored. See ubc_unmap(). As a result ubc_msync() might fail without
the error being propagated back to user space.