Follow recommendations for cdevsw_add()
If index is negative, we start looking for a free slot at the absolute
value of index, instead of starting at 0.

In practice, -1 is unusable, since there are kernel internal devices
that call this function with absolute index values, which will stomp on
free-slot based assignments that happen before them.

12 is considered a safe starting index for Mac OS X 10.5 to 10.7, 24
for OS X 10.8 and later. See bsd/kern/bsd_stubs.c for details.