Remove callbacks for answered tickets
Make sure to remove a ticket's callback when marking it as answered.
Otherwise the ticket will be retained by the callback queue until the
volume is unmounted.

Note: If the file system is marked as dead we don't care about removing
ticket callbacks. All remaining callbacks will be removed automatically
once the volume is unmounted.