Lock parent directory when removing entry
When removing a directory entry HFS and NFS lock the vnode that is to
be removed and its parent directory vnode. However, Apple's fallback
implementation for non-thread-safe 32 bit file systems only locks the
directory entry vnode.

The osxfuse locking mechanism is inspired by Apple's fallback
implementation. While it is not strictly required to lock the parent
directory vnode as well, it seems to be good practice to do so.

Addresses osxfuse/osxfuse#299