Remove cache in case of FUSE_NOTIFY_DELETE
When receiving a FUSE_NOTIFY_DELETE notification the kernel extension
checks if the vnode's nodeid matches the nodeid, that was sent with the
FUSE_NOTIFY_DELETE notification. If they match we will make the vnode
disappear.

If they don't match the kernel extension needs to remove the vnode from
the name cache to prevent race conditions. FUSE_NOTIFY_DELETE should
behave like FUSE_NOTIFY_INVAL_ENTRY, in this case.