Support DEBUG and DEVELOPMENT kernels
We need to know whether a thread is scheduled to be aborted or not.
There is no public API we could use to determine this. So we need to
lookup the relevant field in the kernel-private struct thread. However
the field's offset is different for RELEASE, DEBUG, and DEVELOPMENT
kernels. This commit adds support for non-RELEASE kernels.