Do not post unnecessary file system events
We need to avoid posting unnecessary file system events because they
can have a negative impact on performance. Especially expensive
readdir() calls should be avoided.

If FN_NO_AUTO_NOTIFY is set, no automatic file system events will be
posted for the particular vnode. This is useful when we already know
that we will not be able to detect remote changes reliably.