Do not post unnecessary file system events
Invalidating a vnode's cached attributes essentially means that we
expect the attributes to have changed in user space. However, this also
means that we should not try to be too smart and post automatically
generated file system events for this particular vnode until the
vnode's attributes have been fetched from user space again.